distributed passive master: enable master config change on export (cfg option), introduce mid-master.pl refs #1866
basically, this adds a new config option, and does not harm any existing
installs. if enabled, it will chose to hook mid.pl like normal, but to
further call master specific changes to passivy it further in
mid-master.pl - after that, the icinga config is generated exclusively
for the master from that source, while leaving the hashdump for the
slaves intact.

this is only the first attempt, names may change during further tests.

refs #1866